+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG.md
 
+## Unreleased
+
+### Changed
+
+- Changed multi-user key encapsulation, see [this PR](https://github.com/encryption4all/ibe/pull/10).
+- Changed `irmaseal-curve 0.1.4` to `pg-curve 0.2.0` (includes `bls12_381 0.8`).
+
 ## 0.2.3
 
 ### Added